Extracts navigation logic from AbstractNavtable to a new Navigation class
